p4i integration Integrate 122771 from 3 to main: Get rid of the GL library dependency in Qt (X11 only) by resolving the GL fu's used manually. The GL Visual probing done when the QApplication was constructed is also gone - it's moved to the QGLWidget::renderPixmap() fu. There is now a configure switch (-dlopen-opengl) to enable this, by default the dependencies are still there.
This has been tested both under Linux and IRIX64/32.

Done with Brad and Lars.

[git-p4: depot-paths = "//depot/qt/main/": change = 122785]

+++ src/kernel/qpixmap_x11.cpp
@@ -1389,7 +1389,7 @@ bool QPixmap::convertFromImage( const QImage &img, int conversion_flags )
 	}
 
 	p = newbits;
-	for ( i=0; i<nbytes; i++ ) {		// translate pixels
+	for ( i=0; i< (uint) nbytes; i++ ) {		// translate pixels
 	    *p = pix[*p];
 	    p++;
 	}
